started to debug the tab list result. sadly doesnt work with compact tab list somehow?
